after creating a search using existing patient

+++ database/migrations/2020_11_03_083155_add_version_relations_to_diagnosis.php
@@ -27,7 +27,9 @@ class AddVersionRelationsToDiagnosis extends Migration
     public function down()
     {
         Schema::table('diagnoses', function (Blueprint $table) {
-          $table->dropColumn('version_id');
+          if(Schema::hasColumn('diagnoses', 'version_id')){
+            $table->dropColumn('version_id');
+          }
         });
     }
 }
